Adds stream to docs
